Feature/front gps

See merge request s13-webmobile4-sub1/S13P11A301!84

+++ frontend/app/src/main/java/com/example/bogoargo/data/dto/request/LocationRequestDto.kt
@@ -1,6 +1,8 @@
 package com.example.bogoargo.data.dto.request
 
+import java.math.BigDecimal
+
 data class UserCoordinatesRequest(
-    val latitude: Double,
-    val longitude: Double
+    val latitude: BigDecimal,
+    val longitude: BigDecimal
 )
